fix: 删除beta文件
